+++ src-test/TestUtils.h
@@ -60,7 +60,7 @@ public:
     }
 
     static void extractTransform(Transform & transform, const size_t numTransform, Transform *& outTrans) {
-        ASSERT_TRUE(transform.childrenCount() > numTransform) << " value is " << numTransform;
+        ASSERT_TRUE(transform.childrenNum() > numTransform) << " value is " << numTransform;
         outTrans = static_cast<Transform*>(transform.childAt(numTransform));
     }