+++ src-test/Transformation/TestTransformation-case-1.cpp
@@ -114,9 +114,9 @@ TEST(TestTransform_case1, tr_case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
-    Transform & transformOut3 = transformOut2.createChild(TOTEXT(transformOut3));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
+    Transform & transformOut3 = transformOut2.newChild(TOTEXT(transformOut3));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -241,9 +241,9 @@ TEST(TestTransform_case1, tr_case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
-    Transform & transformOut3 = transformOut2.createChild(TOTEXT(transformOut3));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
+    Transform & transformOut3 = transformOut2.newChild(TOTEXT(transformOut3));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -368,9 +368,9 @@ TEST(TestTransform_case1, tr_case3) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
-    Transform & transformOut3 = transformOut2.createChild(TOTEXT(transformOut3));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
+    Transform & transformOut3 = transformOut2.newChild(TOTEXT(transformOut3));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -461,9 +461,9 @@ TEST(TestTransform_case1, r_case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
-    Transform & transformOut3 = transformOut2.createChild(TOTEXT(transformOut3));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
+    Transform & transformOut3 = transformOut2.newChild(TOTEXT(transformOut3));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -556,9 +556,9 @@ TEST(TestTransform_case1, r_case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
-    Transform & transformOut3 = transformOut2.createChild(TOTEXT(transformOut3));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
+    Transform & transformOut3 = transformOut2.newChild(TOTEXT(transformOut3));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -652,9 +652,9 @@ TEST(TestTransform_case1, t_case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
-    Transform & transformOut3 = transformOut2.createChild(TOTEXT(transformOut3));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
+    Transform & transformOut3 = transformOut2.newChild(TOTEXT(transformOut3));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));