+++ src-test/Transformation/TestTransformation-case-5.cpp
@@ -109,8 +109,8 @@ TEST(DISABLED_TestTransform_case5, tr_case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -222,8 +222,8 @@ TEST(DISABLED_TestTransform_case5, tr_case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -340,8 +340,8 @@ TEST(TestTransform_case5, r_case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -449,8 +449,8 @@ TEST(TestTransform_case5, r_case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -560,8 +560,8 @@ TEST(DISABLED_TestTransform_case5, t_case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -674,8 +674,8 @@ TEST(DISABLED_TestTransform_case5, t_case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -788,8 +788,8 @@ TEST(TestTransform_case5, t_case3) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));