+++ src-test/Transformation/TestTransformationAlgorithm-case-0.cpp
@@ -55,8 +55,8 @@ TEST(TestTransformationAlgorithm_case0, case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
-    Transform & transformOut2 = transformOut1.createChild(TOTEXT(transformOut2));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
+    Transform & transformOut2 = transformOut1.newChild(TOTEXT(transformOut2));
 
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     transformOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -107,7 +107,7 @@ TEST(TestTransformationAlgorithm_case0, case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & transformOut0 = lodOut.transform();
-    Transform & transformOut1 = transformOut0.createChild(TOTEXT(transformOut1));
+    Transform & transformOut1 = transformOut0.newChild(TOTEXT(transformOut1));
     transformOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
 
     mainOut.pMatrix.rotateDegreesY(-90.0f);