+++ src-test/Transformation/TestTransformationAlgorithm-case-6.cpp
@@ -57,8 +57,8 @@ TEST(TestTransformationAlgorithm_case6, case1) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & trOut0 = lodOut.transform();
-    Transform & trOut1 = trOut0.createChild(TOTEXT(trOut1));
-    Transform & trOut2 = trOut1.createChild(TOTEXT(trOut2));
+    Transform & trOut1 = trOut0.newChild(TOTEXT(trOut1));
+    Transform & trOut2 = trOut1.newChild(TOTEXT(trOut2));
 
     trOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     trOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -117,8 +117,8 @@ TEST(TestTransformationAlgorithm_case6, case2) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & trOut0 = lodOut.transform();
-    Transform & trOut1 = trOut0.createChild(TOTEXT(trOut1));
-    Transform & trOut2 = trOut1.createChild(TOTEXT(trOut2));
+    Transform & trOut1 = trOut0.newChild(TOTEXT(trOut1));
+    Transform & trOut2 = trOut1.newChild(TOTEXT(trOut2));
 
     trOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     trOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -178,8 +178,8 @@ TEST(TestTransformationAlgorithm_case6, case3) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & trOut0 = lodOut.transform();
-    Transform & trOut1 = trOut0.createChild(TOTEXT(trOut1));
-    Transform & trOut2 = trOut1.createChild(TOTEXT(trOut2));
+    Transform & trOut1 = trOut0.newChild(TOTEXT(trOut1));
+    Transform & trOut2 = trOut1.newChild(TOTEXT(trOut2));
 
     trOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     trOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));
@@ -243,8 +243,8 @@ TEST(TestTransformationAlgorithm_case6, case4) {
     ObjLodGroup & lodOut = mainOut.addLod();
 
     Transform & trOut0 = lodOut.transform();
-    Transform & trOut1 = trOut0.createChild(TOTEXT(trOut1));
-    Transform & trOut2 = trOut1.createChild(TOTEXT(trOut2));
+    Transform & trOut1 = trOut0.newChild(TOTEXT(trOut1));
+    Transform & trOut2 = trOut1.newChild(TOTEXT(trOut2));
 
     trOut1.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 1"));
     trOut2.addObject(TestUtilsObjMesh::createPyramidTestMesh("mesh 2"));