+++ src/algorithms/InstancingAlg.cpp
@@ -77,7 +77,7 @@ void InstancingAlg::proccessTransform(Transform & transform, bool & outResult) {
     //-------------------------------------------------------------------------
     // children
 
-    const auto chCount = transform.childrenCount();
+    const auto chCount = transform.childrenNum();
     for (Transform::TransformIndex i = 0; i < chCount; ++i) {
         proccessTransform(*static_cast<Transform*>(transform.childAt(i)), outResult);
     }