+++ src/io/writer/ObjWriteAnim.cpp
@@ -94,7 +94,7 @@ bool ObjWriteAnim::printAnimationStart(AbstractWriter & writer, const Transform 
     if (!transform.hasAnim())
         return false;
 
-    if (!transform.hasObjects() && transform.childrenCount() == 0)
+    if (!transform.hasObjects() && transform.childrenNum() == 0)
         return false;
 
     mWriter = &writer;