+++ changelog.md
@@ -10,6 +10,7 @@ StepToSky: X-Plane Obj Library
 - **Changes:** Methods for accessing to the LODs.
 - **Changes:** Changed named light's methods.
 - **Changes:** Changed param light's methods.
+- **Changes:** Changed Transform's methods.
 - **Removed:** ELightNamed and ELightParams. 
                For getting light names use `X-Plane Root/Resources/bitmaps/world/lites/lights.txt`